DOM: Implement <iframe> state-preserving atomic moves
This CL implements iframe-specific logic for state-preserving atomic
moves, as well as some web platform tests and unit tests. We
specifically that iframe documents are not reloaded during a call to
`moveBefore()`, and that the order of frames in `window.frames` remains
untouched. Much of the implementation revolves around how we manage
internal bookkeeping of subframe counts; this bookkeeping is used for
real world behavior in the node insertion/removal path [1], internal
optimizations [2], and Blink logic consistency assertions.

There is still implementation work general to all nodes
(non-iframe-specific) to be done for state-preserving atomic moves
overall, like suppressing mutation events, and tweaking how
MutationObserver reacts to `moveBefore()` calls. That work will be done
in follow-up CLs.

[1]: https://source.chromium.org/chromium/chromium/src/+/main:third_party/blink/renderer/core/dom/container_node.cc;l=834-836;drc=1f0fdce87dc82416abdf7ace3e096e3dbfc7d405

[2]: https://source.chromium.org/chromium/chromium/src/+/main:third_party/blink/renderer/core/dom/child_frame_disconnector.cc;l=21-22;drc=047c7dc4ee1ce908d7fea38ca063fa2f80f92c77

Bug: 40150299
Change-Id: If22ce5f96f8fbe98e126f2e8909aa852e5307bd6