Make containing and sticky rect relative to the scroll container
In the case "flex-direction: column-reverse" or "writing-mode: vertical-rl",
the offset of scrollable overflow rect may be not (0,0),
top or left could be negative.

If the sticky container is not the same with scroll container,
then everything is relative to scroll container,
so sticky item works fine.

If sticky container is the same with scroll container, then containing
block rect and sticky box rect could have negative offset because they
are based on scrollable overflow rect.

But if ignored border offset, when scroll position is 0,
sticky container is at 0,0 too. It only move to negative scrollable
overflow rect when scrollbar scrolled to bottom.

So the initial scroll container relative containing block rect should
not base on the scrollable overflow rect. We must adjust containing block
rect based on ScrollOrigin to move it to correct offset.
The sticky box rect is the same.

Bug: 331753412
Change-Id: I877b73afc458cd0de0ecfb618e33bc95118380a1