Make room for @page borders and padding in layout.
Page borders and padding shouldn't be affected by layout scaling. Since
the page border box lives in the coordinate system of layout, the values
need to be scaled down when computing style.

The insets caused by @page borders and padding will naturally shrink the
size available to the page content area. As such
CalculateChildAvailableSize() needs an update now. The page container
and border boxes *are* anonymous blocks, but their block-size should be
allowed to be adjusted nevertheless. Only prevent block-size adjustment
if it's indefinite (which is never the case for page containers or page
border boxes).

Bug: 40286153
Change-Id: Ie701aa65cd5f23820d3a8de3e40958342a106174